Splashscreen

+++ Game/src/Layers/townlayer.cpp
@@ -294,7 +294,7 @@ namespace Game
     // Randomly toggle light on and off between range 1 to 2 seconds only in the other scene, but of course when added this can be added to the other scene as well
     if (town_version == "assets/saves/town_v7_2.flxscene")
     {
-      static float light_timer = rand() % 1 + 1;
+      static float light_timer = (float)(rand() % 1 + 1);
       light_timer -= Application::GetCurrentWindow()->GetFramerateController().GetDeltaTime();
       if (light_timer <= 0)
       {
@@ -302,7 +302,7 @@ namespace Game
         FlexECS::Scene::GetEntityByName("Flicker").GetComponent<Transform>()->is_active = 
           !FlexECS::Scene::GetEntityByName("Flicker").GetComponent<Transform>()->is_active;
       
-        light_timer = rand() % 1 + 1;
+        light_timer = (float)(rand() % 1 + 1);
       }
     }
